+++ WeatherForecast/WeatherForecast/Controllers/WeatherListViewController.swift
@@ -35,7 +35,9 @@ class WeatherListViewController: UIViewController {
     }
     
     private func layoutWeatherListView() {
-        weatherListView = .init(weatherInfo: weatherInfo, imageService: ImageService())
+        weatherListView = .init(weatherInfo: weatherInfo,
+                                imageService: ImageService(),
+                                jsonService: JsonService())
         view.addSubview(weatherListView)
         weatherListView.translatesAutoresizingMaskIntoConstraints = false
         weatherListView.delegate = self
