+++ WeatherForecast/WeatherForecast/Networks/Serviceable/JsonServiceable.swift
@@ -1,0 +1,12 @@
+//
+//  JsonServiceable.swift
+//  WeatherForecast
+//
+//  Created by 김창규 on 1/30/24.
+//
+
+import Foundation
+
+protocol JsonServiceable {
+    func fetchWeatherJSON() -> WeatherJSON?
+}
